win32-dont-attempt-restart.diff: Don't attempt to restart OOo after crash
n#524250

Don't attempt to restart soffice.bin immediately if it crashes.  At
such a restart, if the document recovery dialog is cancelled, an
empty Writer document is opened even if the user was editing some
other kind of document. Some customers are annoyed by this. And the
automatic restart doesn't work on Linux anyway, so for consistency
it doesn't hurt if we force it not to work on Windows, either.