prevent symbols cflags from overriding debug cflags
The use case here is configuring with --enable-symbols=small (which
enables -g1 for gcc), and then doing 'dmake debug=true', which
enables -g, however the -g1 is later in the command line, so it
takes precedence. And I don't feel like trying to redo the whole
ordering of gcc command lines everywhere, while this looks safe.