fixed an i18n problem
Constructing "string + ordinal number" terms programmatically is
an i18n issue, because in some languages (e.g. in Hungarian)
ordinal number should precede the string. The patch solves
http://bug.openscope.org/browse/OOO-85 reported by a Hungarian
localizer ("Oldal 1" should be "1. oldal"). There are more
occurences of this construction in the code of LibreOffice
but this was highly visible and it is correct in Hungarian Excel.