When launching the ref input window, we need to select the initial string.
Launching the reference input window from the formula wizard dialog did not
set initial selection over the whole reference string.  For instance, type
=ABS(A1) in a cell, open up the formula wizard dialog, then click the
reference button to launch the reference input which has 'A1' inside, this
string.  Because 'A1' was not selected, when you click on another cell it
would not overwrite the old reference string.

This commit fixes it.