Allow the dropdown list of a combo box to be scrollable. (fdo#31710)
This change makes the dropdown list of a combo box control to be
horizontally scrollable in case the longest text in the list box
is wider than the maximum allowed width of the dropdown list.

Also, previously the dropdown list would become wider than the
desktop width when the longest text exceeded the desktop width.
This change caps the max width of the dropdown list to be the width
of the desktop width.