disable multiple migrations via MIGRATED stamp file
If we fail to auto-migrate, this can cause lots of pain - since people
are unaware of where the migration occurs from. Thus - before we start
migrating, we create a 'MIGRATED' stamp file in the source of that, and
if this is present next migration, we simply don't migrate - fixed for
Mechtilde.