exile some download-dependent configure test into a post_download step
on windows, the configure check that some dll/exe are present.
these are in fact automatically downloaded by the ./download script.
but that script is run _after_ ./configure.
So the tests are postponed and run in a post_download script that is
invoked at the end of the download script.