v3 of the configure.in/scp2 patch to allow selecting themes
DOH, here comes v3, I left a debug output in configure.in in v2, otherwise no change. Sorry

Sebastian

>From 6c379488250f5a3349dfc2d805f70a0f084fe07d Mon Sep 17 00:00:00 2001
From: Sebastian Spaeth <Sebastian@SSpaeth.de>
Date: Thu, 9 Dec 2010 12:23:55 +0100
Subject: [PATCHv3] Introduce an --enable-theme="theme1 theme2..." option

By default, we include all defaults as before, but if given, we can
limit the number of included themes with e.g. --enable-theme="oxygen
tango". This is passed through set_soenv and sets the environment
variable ENABLE_THEMES="...." which can be used by the packimages
makefile in the postprocess repository. It also adds defines in the
form of THEME_OXYGEN that are being used in the scp2 module to decide
if we should install the theme file or not.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>