make public
The SdrDragEntry class is "leaked out" through the functions of
SVX_DLLPUBLIC SdrDragMethod that use its maSdrDragEntries member (of
type std::vector<SdrDragEntry*>). That breaks build with debug version
of libstdc++ (with gcc 4.5, at least) in sd and chart2 because linker
cannot find typeinfo for SdrDragEntry.

Alternative would be to move the definitions of these functions to
source/svdraw/svddrgmt.cxx .