Let's hide this ugliness inside the remove-erase idiom.
So, since I can't rely on the iterator returned from vector::erase(),
and I can't keep the iterator of the erased element (which is probably
invalidated after the erase call), I have to erase elements by
other means.  Luckily this remove-erase idiom doesn't cause a segfault,
and appears to do the right thing.