Avoid unnecessary PATH manipulation for exes located in OUTDIR/bin
svidl.exe, transex3.exe and xsltproc.exe are themselves in
${OUTDIR}/bin, so no need to add that directory to PATH when running
them. They will find the DLLs they need from there automatically
anyway.

Besides, the PATH manipulation was wrong as OUTDIR here is a Win32
pathname, starting with a drive letter and colon, while PATH is the
Cygwin one that contains Cygwin pathnames with no drive letters and
with colon as the search path separator. Thus a cygpath -u
transformation of OUTDIR would have been needed to actually make it do
what was intended.