add presumably missing ReleaseSolarMutex() call
Noticed by Julien Nabet <serval2412@yahoo.fr>. It seems commit
b58863e8ddc923fa737d064b873b8099a24a8c51 reorganized the code
and removed this call. Not quite sure if it's needed as it has
apparently worked fine for a while, but this should match the code
before that commit (except for the pointless TryToAcquire() ).