fdo#36406: Let's not use invalidated iterators.
std::set::erase(iterator) call invalidates the iterator of the erased
element.  We better not use it after the erase() call.  Since the number
of manual breaks should not be high enough to cause a performance
bottleneck under normal usage, a safer linear copy should just be fine.

Now, I'm not sure if this indeed is the cause of the erroneous amount
of manual breaks in the test document, but I didn't see any other likely
spot.