Very early hacking on MinGW cross-compilation
We need to differentiate between compiling for Windows, and building
*on* Windows (which always means on Cygwin, actually). Fixed some of
those cases to use the proper test. cygpath should be used only when
building on Cygwin, naturally.

If attempting a MinGW cross-compilation it still doesn't even pass the
configure script without failing. Many hard problems have not been
solved.

Cygwin's gcc -mno-cygwin is generally considered a broken concepts, so
binning support for that. Instead, if on Windows but not using MSVC
one should use the Cygwin-based MinGW cross-compiler. That case should
probably not be considered as cross-compilation, though.