rework of fix for fdo#35826
The problem stems from how and when we get notification ( via EndURL ) of the termination of the hyperlink. On investigation it seems that if we get EndURL before RunText then it's the previous text run that is enclosed in the hyperlink
if we get endURL after Runtext then it's *this* run that needs to have a following hyperlink close.