MinGW hacks, we get a bit further now
I am mostly ignoring the possibility to build natively on Windows
using MinGW. I don't see any reasoon to bother supporting that. If we
want to build with MinGW, we want to do it as cross-compilation.

I am not intentionally removing OOo-originated native MinGW
compilation stuff. Yet. But will soon...