Use the sweep flag to determine the drawing direction
The previous solution was giving somehow acceptable results
in the majority of cases, but was pretty broken
with corner cases where the arc is exactly half of the whole ellipse and where
the large arc flag is largely irrelevant.
If the sweep flag is equal to 0 drawing goes in the sense of decreasing angles.
If it is set, drawing goes in the sense of increasing angles.