Allow use of $ACLOCAL to override the name aclocal
Useful for instance when only aclocal-1.10 exists. But, systems that
are broken like that will have lots of more problems, too. It is just
a (sad?) fact that building LO on anything except those platforms that
people work on constantly will require lots of tweaks. Many of the
assumptions in various solenv .mk files for rare platforms are rather
outdated and/or arbitrary, assuming a certain mix of more or less
"official" additional tools.