Header/Footer separators drawing moved to the SwPageFrm level
Drawing the header / footer separator in the SwHeadFootFrm was needing
a call to InvalidateWindows to have a clean drawing... but this
introduced some too frequent redrawing of the headers and footers thus
giving cursor blinking and useless cycles used.

The new drawing is done by the SwPageFrm in the same way than the grid
drawing. It also has the advantage than the separator is now drawn on
top of everything else in the page.