Replace List with std::vector< INetMIMEMessage* >
I also removed "sal_uIntPtr nNumChildren;". The use was question able at
best. Its count could be off because AttachChild() set it to size() if the
child was added; however, if the child wasn't added the person doing the
AttachChild() was also incrementing the count.

Related to the removal of nNumChildren, I also removed the GetChildCount()
& SetChildCount() methods because they where only used around the broken
AttachChild() code.