Correctly use adjusted range strings for range name import.
ScXMLConverter::ParseFormula() was called on the range name string
to correctly fix range string such as '$Sheet1.$A$1:.$A$3' to
'$Sheet1.$A$1:$A$3' (remove the dot from the end cell address).  But
we weren't using the corrected string when creating a new range name
instance during the ODS import.  This commit fixes it.