Remove superfluous empty line
The comment for gb_Library_ILIBFILENAMES in the WNTGCC case doesn't
make much sense for us, though, as I think we want to support WNTGCC
only for cross-compilation, and thus we don't have access to the PSDK
("Platform" SDK, with which they mean the Windows SDK). I don't see
any point in using MinGW natively on Windows to build LibreOffice. If
you build locally, use MSVC.